+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/PartyManagementServiceIT.scala
@@ -21,7 +21,7 @@ import scalaz.Tag
 import scalaz.syntax.tag.ToTagOps
 import java.util.regex.Pattern
 
-import com.daml.ledger.api.testtool.suites.v1_8.object_meta.ObjectMetaTestsForPartyManagementService
+import com.daml.ledger.api.testtool.suites.v1_8.objectmeta.PartyManagementObjectMetaTests
 import com.daml.ledger.api.v1.admin.object_meta.ObjectMeta
 import com.daml.ledger.client.binding.Primitive
 
@@ -31,7 +31,7 @@ final class PartyManagementServiceIT
     extends LedgerTestSuite
     with PartyManagementItUtils
     with PartyManagementServiceUpdateRpcTests
-    with ObjectMetaTestsForPartyManagementService {
+    with PartyManagementObjectMetaTests {
 
   test(
     "PMNonEmptyParticipantID",
