+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/UserManagementServiceIT.scala
@@ -11,7 +11,7 @@ import com.daml.error.utils.ErrorDetails
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
 import com.daml.ledger.api.testtool.infrastructure.Assertions.{assertEquals, _}
 import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
-import com.daml.ledger.api.testtool.suites.v1_8.object_meta.ObjectMetaTestsForUserManagementService
+import com.daml.ledger.api.testtool.suites.v1_8.objectmeta.UserManagementObjectMetaTests
 import com.daml.ledger.api.v1.admin.object_meta.ObjectMeta
 import com.daml.ledger.api.v1.admin.user_management_service.{
   CreateUserRequest,
@@ -46,7 +46,7 @@ final class UserManagementServiceIT
     with UserManagementServiceITUtils
     with UserManagementServiceUpdateRpcTests
     with UserManagementServiceUpdatePrimitivePropertiesTests
-    with ObjectMetaTestsForUserManagementService {
+    with UserManagementObjectMetaTests {
 
   private val adminPermission =
     Permission(Permission.Kind.ParticipantAdmin(Permission.ParticipantAdmin()))
