+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/objectmeta/PartyManagementObjectMetaTests.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.testtool.suites.v1_8.object_meta
+package com.daml.ledger.api.testtool.suites.v1_8.objectmeta
 
 import com.daml.error.definitions.LedgerApiErrors
 import com.daml.ledger.api.testtool.infrastructure.ExpectedErrorDescription
@@ -16,34 +16,34 @@ import com.daml.ledger.api.v1.admin.party_management_service.{
 
 import scala.concurrent.{ExecutionContext, Future}
 
-trait ObjectMetaTestsForPartyManagementService extends ObjectMetaTests with ObjectMetaTestsBase {
+trait PartyManagementObjectMetaTests extends ObjectMetaTests {
   self: PartyManagementServiceIT =>
 
   type Resource = PartyDetails
   type ResourceId = String
 
-  override private[object_meta] def getId(resource: Resource): ResourceId = resource.party
+  override private[objectmeta] def getId(resource: Resource): ResourceId = resource.party
 
-  override private[object_meta] def annotationsUpdateRequestFieldPath: String =
+  override private[objectmeta] def annotationsUpdateRequestFieldPath: String =
     "party_details.local_metadata.annotations"
 
-  override private[object_meta] def resourceVersionUpdatePath: String =
+  override private[objectmeta] def resourceVersionUpdatePath: String =
     "local_metadata.resource_version"
 
-  override private[object_meta] def annotationsUpdatePath: String =
+  override private[objectmeta] def annotationsUpdatePath: String =
     "local_metadata.annotations"
 
-  override private[object_meta] def annotationsShortUpdatePath = "local_metadata"
+  override private[objectmeta] def annotationsShortUpdatePath = "local_metadata"
 
-  override private[object_meta] def resourceIdPath = "party"
+  override private[objectmeta] def resourceIdPath = "party"
 
-  override private[object_meta] def extractAnnotations(resource: Resource): Map[String, String] =
+  override private[objectmeta] def extractAnnotations(resource: Resource): Map[String, String] =
     resource.getLocalMetadata.annotations
 
-  override private[object_meta] def extractMetadata(resource: Resource): ObjectMeta =
+  override private[objectmeta] def extractMetadata(resource: Resource): ObjectMeta =
     resource.getLocalMetadata
 
-  override private[object_meta] def testWithFreshResource(
+  override private[objectmeta] def testWithFreshResource(
       shortIdentifier: String,
       description: String,
   )(
@@ -58,7 +58,7 @@ trait ObjectMetaTestsForPartyManagementService extends ObjectMetaTests with Obje
     )(annotations = annotations)(body)
   }
 
-  override private[object_meta] def testWithoutResource(
+  override private[objectmeta] def testWithoutResource(
       shortIdentifier: String,
       description: String,
   )(
@@ -71,7 +71,7 @@ trait ObjectMetaTestsForPartyManagementService extends ObjectMetaTests with Obje
     )(body)
   }
 
-  override private[object_meta] def createResourceWithAnnotations(
+  override private[objectmeta] def createResourceWithAnnotations(
       annotations: Map[String, String]
   )(implicit ec: ExecutionContext, ledger: ParticipantTestContext): Future[Map[String, String]] = {
     val req = AllocatePartyRequest(localMetadata = Some(ObjectMeta(annotations = annotations)))
@@ -80,7 +80,7 @@ trait ObjectMetaTestsForPartyManagementService extends ObjectMetaTests with Obje
       .map(extractUpdatedAnnotations)
   }
 
-  override private[object_meta] def fetchNewestAnnotations(
+  override private[objectmeta] def fetchNewestAnnotations(
       id: ResourceId
   )(implicit ec: ExecutionContext, ledger: ParticipantTestContext): Future[Map[String, String]] = {
     ledger
@@ -88,7 +88,7 @@ trait ObjectMetaTestsForPartyManagementService extends ObjectMetaTests with Obje
       .map(_.partyDetails.head.getLocalMetadata.annotations)
   }
 
-  override private[object_meta] def update(
+  override private[objectmeta] def update(
       id: ResourceId,
       annotations: Map[String, String],
       updatePaths: Seq[String],
@@ -105,7 +105,7 @@ trait ObjectMetaTestsForPartyManagementService extends ObjectMetaTests with Obje
       .map(_.getPartyDetails.getLocalMetadata)
   }
 
-  override private[object_meta] def concurrentUserUpdateDetectedErrorDescription(
+  override private[objectmeta] def concurrentUserUpdateDetectedErrorDescription(
       id: ResourceId
   ): ExpectedErrorDescription = ExpectedErrorDescription(
     errorCode = LedgerApiErrors.Admin.PartyManagement.ConcurrentPartyDetailsUpdateDetected,
@@ -114,7 +114,7 @@ trait ObjectMetaTestsForPartyManagementService extends ObjectMetaTests with Obje
     ),
   )
 
-  override private[object_meta] def invalidUpdateRequestErrorDescription(
+  override private[objectmeta] def invalidUpdateRequestErrorDescription(
       id: ResourceId,
       errorMessageSuffix: String,
   ): ExpectedErrorDescription = ExpectedErrorDescription(
