+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/objectmeta/UserManagementObjectMetaTests.scala
@@ -4,7 +4,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.testtool.suites.v1_8.object_meta
+package com.daml.ledger.api.testtool.suites.v1_8.objectmeta
 
 import com.daml.error.definitions.LedgerApiErrors
 import com.daml.ledger.api.testtool.infrastructure.ExpectedErrorDescription
@@ -19,33 +19,33 @@ import com.daml.ledger.api.v1.admin.user_management_service.{
 
 import scala.concurrent.{ExecutionContext, Future}
 
-trait ObjectMetaTestsForUserManagementService extends ObjectMetaTests with ObjectMetaTestsBase {
+trait UserManagementObjectMetaTests extends ObjectMetaTests with ObjectMetaTestsBase {
   self: UserManagementServiceIT =>
 
   type Resource = User
   type ResourceId = String
 
-  override private[object_meta] def getId(resource: Resource): ResourceId = resource.id
+  override private[objectmeta] def getId(resource: Resource): ResourceId = resource.id
 
-  override private[object_meta] def annotationsUpdateRequestFieldPath: String =
+  override private[objectmeta] def annotationsUpdateRequestFieldPath: String =
     "user.metadata.annotations"
 
-  override private[object_meta] def annotationsUpdatePath: String =
+  override private[objectmeta] def annotationsUpdatePath: String =
     "metadata.annotations"
 
-  override private[object_meta] def annotationsShortUpdatePath = "metadata"
+  override private[objectmeta] def annotationsShortUpdatePath = "metadata"
 
-  override private[object_meta] def resourceVersionUpdatePath = "metadata.resource_version"
+  override private[objectmeta] def resourceVersionUpdatePath = "metadata.resource_version"
 
-  override private[object_meta] def resourceIdPath = "id"
+  override private[objectmeta] def resourceIdPath = "id"
 
-  override private[object_meta] def extractAnnotations(resource: Resource): Map[String, String] =
+  override private[objectmeta] def extractAnnotations(resource: Resource): Map[String, String] =
     resource.getMetadata.annotations
 
-  override private[object_meta] def extractMetadata(resource: Resource): ObjectMeta =
+  override private[objectmeta] def extractMetadata(resource: Resource): ObjectMeta =
     resource.getMetadata
 
-  override private[object_meta] def testWithFreshResource(
+  override private[objectmeta] def testWithFreshResource(
       shortIdentifier: String,
       description: String,
   )(
@@ -66,7 +66,7 @@ trait ObjectMetaTestsForUserManagementService extends ObjectMetaTests with Objec
     })
   }
 
-  override private[object_meta] def testWithoutResource(
+  override private[objectmeta] def testWithoutResource(
       shortIdentifier: String,
       description: String,
   )(
@@ -81,7 +81,7 @@ trait ObjectMetaTestsForUserManagementService extends ObjectMetaTests with Objec
     })
   }
 
-  override private[object_meta] def createResourceWithAnnotations(
+  override private[objectmeta] def createResourceWithAnnotations(
       annotations: Map[String, String]
   )(implicit ec: ExecutionContext, ledger: ParticipantTestContext): Future[Map[String, String]] = {
     val userId = ledger.nextUserId()
@@ -98,7 +98,7 @@ trait ObjectMetaTestsForUserManagementService extends ObjectMetaTests with Objec
       .map(extractAnnotations)
   }
 
-  override private[object_meta] def fetchNewestAnnotations(
+  override private[objectmeta] def fetchNewestAnnotations(
       id: ResourceId
   )(implicit ec: ExecutionContext, ledger: ParticipantTestContext): Future[Map[String, String]] = {
     ledger.userManagement
@@ -106,7 +106,7 @@ trait ObjectMetaTestsForUserManagementService extends ObjectMetaTests with Objec
       .map(_.user.get.getMetadata.annotations)
   }
 
-  override private[object_meta] def update(
+  override private[objectmeta] def update(
       id: ResourceId,
       annotations: Map[String, String],
       updatePaths: Seq[String],
@@ -123,7 +123,7 @@ trait ObjectMetaTestsForUserManagementService extends ObjectMetaTests with Objec
       .map(_.getUser.getMetadata)
   }
 
-  override private[object_meta] def concurrentUserUpdateDetectedErrorDescription(
+  override private[objectmeta] def concurrentUserUpdateDetectedErrorDescription(
       id: ResourceId
   ): ExpectedErrorDescription = ExpectedErrorDescription(
     errorCode = LedgerApiErrors.Admin.UserManagement.ConcurrentUserUpdateDetected,
@@ -132,7 +132,7 @@ trait ObjectMetaTestsForUserManagementService extends ObjectMetaTests with Objec
     ),
   )
 
-  override private[object_meta] def invalidUpdateRequestErrorDescription(
+  override private[objectmeta] def invalidUpdateRequestErrorDescription(
       id: ResourceId,
       errorMessageSuffix: String,
   ): ExpectedErrorDescription = ExpectedErrorDescription(
